Kynn up code Promotion and Customer

+++ PhanMemQuanLyBanQuanAo/src/main/java/com/Models/Promotion.java
@@ -30,7 +30,7 @@ public class Promotion implements Serializable {
     private String name;
     
     @Column(name = "decreaseNumber")
-    private Float DecreaseNumber;
+    private int DecreaseNumber;
 
     @Column(name = "StartDay")
     private Date startDay;
@@ -44,7 +44,7 @@ public class Promotion implements Serializable {
     @OneToMany(mappedBy = "promotion", fetch = FetchType.LAZY)
     private List<Bill> listBill;
 
-    public Promotion(String code, String name, Float DecreaseNumber, Date startDay, Date endDay, int status) {
+    public Promotion(String code, String name, int DecreaseNumber, Date startDay, Date endDay, int status) {
         this.code = code;
         this.name = name;
         this.DecreaseNumber = DecreaseNumber;
